feat: upgrade native audio codecs, modernize build system, and add multi-platform GitHub Actions workflow

+++ natives/connector/mpg123.c
@@ -27,11 +27,11 @@ CONNECTOR_EXPORT jlong JNICALL Java_com_sedmelluq_discord_lavaplayer_natives_mp3
 		return 0;
 	}
 	
-	return (jlong) handle;
+	return (jlong)(uintptr_t) handle;
 }
 
 CONNECTOR_EXPORT void JNICALL Java_com_sedmelluq_discord_lavaplayer_natives_mp3_Mp3DecoderLibrary_destroy(JNIEnv *jni, jobject me, jlong instance) {
-	void* handle = (void*) instance;
+	void* handle = (void*)(uintptr_t) instance;
 	
 	if (handle != NULL) {
 		mpg123_close(handle);
@@ -50,7 +50,7 @@ CONNECTOR_EXPORT jint JNICALL Java_com_sedmelluq_discord_lavaplayer_natives_mp3_
 	unsigned char* output = (*jni)->GetDirectBufferAddress(jni, direct_output);
 	size_t used_bytes = 0;
 	
-	int result = mpg123_decode((void*) instance, input, (size_t) input_length, output, (size_t) output_length, &used_bytes);
+	int result = mpg123_decode((void*)(uintptr_t) instance, input, (size_t) input_length, output, (size_t) output_length, &used_bytes);
 	
 	if (result != 0) {
 		if (result > 0) {
